requirements.txt
#THIS is the requirements as of 21 February 2024 after Q&a
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
backoff==2.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.2
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
fastavro==1.9.3
frozenlist==1.4.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.11.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.339
langchain-community==0.0.12
langchain-core==0.1.10
langchain-experimental==0.0.42
langchain-openai==0.0.2.post1
langsmith==0.0.65
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
numpy==1.26.2
openai==1.7.2
packaging==23.2
pandas==2.1.4
pip==23.2.1
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
djangorestframework-simplejwt==5.3.1
rsa==4.9
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.2
zipp==3.17.0